refactor(db): 重构数据库写入流程和关闭逻辑
- 新增 closer 结构体，用于统一管理数据库关闭流程
- 重构 handleWriteCh 函数，优化写入流程和关闭逻辑
- 修改 Close 函数，确保所有组件正确关闭
- 调整 compact 相关逻辑，增加并发控制和错误处理
- 优化 initVlog 函数，简化值日志初始化流程

+++ src/model/Key.go
@@ -8,20 +8,6 @@ import (
 	errors "trainKv/common"
 )
 
-//func CompareKey(key1, key2 []byte) int {
-//	errors.CondPanic(len(key1) <= 8 || len(key2) <= 8,
-//		fmt.Errorf("CompareKeys:key1:%s len(key1):%d < 8 || key2:%s len(key2):%d < 8", string(key1), len(key1), string(key2), len(key2)))
-//	if compare := bytes.Compare(key1[0:len(key1)-8], key2[0:len(key2)-8]); compare != 0 {
-//		return compare
-//	}
-//	// key 相同的情况下, 比较后面的 时间戳; 但是会返回 cmp == 0 吗? 时间戳会相同吗? 不会
-//	return bytes.Compare(key1[len(key1)-8:], key2[len(key2)-8:])
-//}
-
-//func CompareBaseKeyNoTs(blockBaseKey, userKey []byte) int {
-//	return bytes.Compare(blockBaseKey[:], userKey[:len(userKey)-8])
-//}
-
 // CompareKeyNoTs skipList.key()  table.biggestKey()  block.baseKey()
 func CompareKeyNoTs(key1, key2 []byte) int {
 	errors.CondPanic(len(key1) <= 8 || len(key2) <= 8, fmt.Errorf("%s,%s < 8", string(key1), string(key2)))
@@ -37,12 +23,12 @@ func CompareKeyWithTs(key1, key2 []byte) int {
 	return bytes.Compare(key1Version, key2Version)
 }
 
-func ParseTsVersion(key []byte) uint64 {
+func ParseTsVersion(key []byte) int64 {
 	if len(key) <= 8 {
 		return 0
 	}
 	timestamp := binary.BigEndian.Uint64(key[len(key)-8:])
-	return timestamp
+	return int64(timestamp)
 }
 
 // ParseKey parses the actual key from the key bytes.
